+++ src/main/java/com/polar_moviechart/userservice/domain/controller/secureapi/AccessTokenDto.java
@@ -1,4 +1,4 @@
-package com.polar_moviechart.userservice.domain.controller;
+package com.polar_moviechart.userservice.domain.controller.secureapi;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
